Add basic Ping-Pong functionality
