+++ MomCare+Watch/Common/WatchConnector+WatchOS.swift
@@ -12,7 +12,10 @@ extension WatchConnector {
     func session(_ session: WCSession, didReceiveMessage message: [String: Any], replyHandler: @escaping ([String: Any]) -> Void) {
         if let text = message["message"] as? String, text == "ping" {
             replyHandler(["message": "pong"])
-            WatchObserver.shared.pongReceived = true
+            DispatchQueue.main.async {
+                WatchObserver.shared.pongReceived = true
+            }
+            Utils.createNotification(title: "MomCare+", body: "Received ping from MomCare+")
         }
     }
 }
