+++ MomCare/Common/CacheHandler/CacheHandler.swift
@@ -1,0 +1,70 @@
+//
+//  CacheHandler.swift
+//  MomCare
+//
+//  Created by RITIK RANJAN on 28/05/25.
+//
+
+import Foundation
+import UIKit
+import OSLog
+import RealmSwift
+
+private let logger: os.Logger = .init(subsystem: "com.MomCare.DataHandler", category: "Cache")
+
+class Entity<T> {
+
+    // MARK: Lifecycle
+
+    init(value: T? = nil, expiration: Date? = nil) {
+        self.value = value
+        self.expiration = expiration ?? Date().addingTimeInterval(60 * 60 * 24)
+    }
+
+    // MARK: Internal
+
+    var value: T?
+    var expiration: Date?
+
+    var isExpired: Bool {
+        guard let expiration else {
+            return false
+        }
+
+        return Date() > expiration
+    }
+
+}
+
+@MainActor
+class CacheHandler {
+
+    // MARK: Public
+
+    public private(set) var cache: NSCache<AnyObject, AnyObject> = .init()
+
+    // MARK: Internal
+
+    static let shared: CacheHandler = .init()
+
+    func set<T>(_ value: T, forKey key: String, expiration: Date? = nil) {
+        let entity = Entity<T>(value: value, expiration: expiration)
+        cache.setObject(entity as AnyObject, forKey: key as AnyObject)
+        logger.debug("Set value for key: \(key)")
+    }
+
+    func get<T>(forKey key: String, expiration: Date? = nil) -> T? {
+        if let entity = cache.object(forKey: key as AnyObject) as? Entity<T> {
+            logger.debug("Retrieved value for key: \(key)")
+            if entity.isExpired {
+                logger.debug("Value for key: \(key) is expired, removing from cache")
+                cache.removeObject(forKey: key as AnyObject)
+            } else {
+                return entity.value
+            }
+        }
+        logger.debug("No value found for key: \(key)")
+        return nil
+    }
+
+}
