+++ MomCare/Common/WatchConnector+iOS.swift
@@ -7,6 +7,11 @@
 
 import WatchConnectivity
 
-#if os(iOS)
-extension WatchConnector {}
-#endif
+extension WatchConnector {
+    func session(_ session: WCSession, didReceiveMessage message: [String: Any], replyHandler: @escaping ([String: Any]) -> Void) {
+        if let text = message["message"] as? String, text == "ping" {
+            replyHandler(["message": "pong"])
+            Utils.createNotification(title: "MomCare+ Watch", body: "Received ping from watch")
+        }
+    }
+}
